fix(lint): only write files when lint rules apply
Don't modify files during `lint --fix` that do not match any lint
warnings.

As a silly example, consider this badly formatted schema that won't
trigger any linting warning:

```json
  {
             :
   http://json-schema.org/draft-06/schema#,    type    : string
}
```

If we run this over `jsonschema lint --fix`, the CLI will prettify
(and thus change the formatting) even though no lint rule applied.
While it is not very harmful, this tends to create a lot of
unnecessary noise when applying the linter on a large amount of files.

Co-Authored-By: Juan Cruz Viotti <jv@jviotti.com>